requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram-dialog==2.0.0b16
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==3.0.1
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.9
MarkupSafe==2.1.2
multidict==6.0.4
pydantic==1.10.5
SQLAlchemy==2.0.4
typing_extensions==4.5.0
yarl==1.8.2
python-dotenv~=1.0.0

pytest~=7.3.1
